+++ src/snowflake/snowpark/modin/plugin/_internal/apply_utils.py
@@ -25,6 +25,7 @@ from snowflake.snowpark.modin.plugin._internal.ordered_dataframe import (
 )
 from snowflake.snowpark.modin.plugin._internal.utils import (
     TempObjectType,
+    generate_snowflake_quoted_identifiers_helper,
     parse_object_construct_snowflake_quoted_identifier_and_extract_pandas_label,
     parse_snowflake_object_construct_identifier_to_map,
 )
@@ -47,11 +48,6 @@ from snowflake.snowpark.window import Window
 
 APPLY_LABEL_COLUMN_QUOTED_IDENTIFIER = '"LABEL"'
 APPLY_VALUE_COLUMN_QUOTED_IDENTIFIER = '"VALUE"'
-APPLY_ORIGINAL_ROW_POSITION_COLUMN_QUOTED_IDENTIFIER = '"ORIGINAL_ROW_POSITION"'
-APPLY_ROW_POSITION_WITHIN_GROUP_COLUMN_QUOTED_IDENTIFIER = '"ROW_POSITION_WITHIN_GROUP"'
-APPLY_FIRST_GROUP_KEY_OCCURRENCE_POSITION_QUOTED_IDENTIFIER = (
-    '"FIRST_GROUP_KEY_OCCURRENCE_POSITION"'
-)
 
 # Default partition size to use when applying a UDTF. A higher value results in less parallelism, less contention and higher batching.
 DEFAULT_UDTF_PARTITION_SIZE = 1000
@@ -387,6 +383,7 @@ def create_udtf_for_groupby_apply(
     session: Session,
     series_groupby: bool,
     by_types: list[DataType],
+    existing_identifiers: list[str],
 ) -> UserDefinedTableFunction:
     """
     Create a UDTF from the Python function for groupby.apply.
@@ -475,14 +472,15 @@ def create_udtf_for_groupby_apply(
     input_index_column_types: Types of the input dataframe's index columns
     series_groupby: Whether we are performing a SeriesGroupBy.apply() instead of DataFrameGroupBy.apply()
     by_types: The snowflake types of the by columns.
+    existing_identifiers: List of existing column identifiers; these are omitted when creating new column identifiers.
 
     Returns
     -------
     A UDTF that will apply the provided function to a group and return a
-    dataframe representing all of the data and metadata of the result.
+    dataframe representing all the data and metadata of the result.
     """
 
-    # Get the length of this list outside of the vUDTF function because the vUDTF
+    # Get the length of this list outside the vUDTF function because the vUDTF
     # doesn't have access to the Snowpark module, which defines these types.
     num_by = len(by_types)
 
@@ -502,7 +500,7 @@ def create_udtf_for_groupby_apply(
             """
             current_column_position = 0
 
-            # First column is row position. Save it for later.
+            # The first column is row position. Save it for later.
             row_position_column_number = 0
             row_positions = df.iloc[:, row_position_column_number]
             current_column_position = row_position_column_number + 1
@@ -614,17 +612,26 @@ def create_udtf_for_groupby_apply(
         # ...then the data columns for the input dataframe or series.
         *input_data_column_types,
     ]
+
+    col_labels = [
+        "LABEL",
+        "ROW_POSITION_WITHIN_GROUP",
+        "VALUE",
+        "ORIGINAL_ROW_POSITION",
+        "APPLY_FIRST_GROUP_KEY_OCCURRENCE_POSITION",
+    ]
+    # Generate new column identifiers for all required UDTF columns with the helper below to prevent collisions in
+    # column identifiers.
+    col_names = generate_snowflake_quoted_identifiers_helper(
+        pandas_labels=col_labels,
+        excluded=existing_identifiers,
+        wrap_double_underscore=False,
+    )
     return udtf(
         ApplyFunc,
         output_schema=PandasDataFrameType(
             [StringType(), IntegerType(), VariantType(), IntegerType(), IntegerType()],
-            [
-                APPLY_LABEL_COLUMN_QUOTED_IDENTIFIER,
-                APPLY_ROW_POSITION_WITHIN_GROUP_COLUMN_QUOTED_IDENTIFIER,
-                APPLY_VALUE_COLUMN_QUOTED_IDENTIFIER,
-                APPLY_ORIGINAL_ROW_POSITION_COLUMN_QUOTED_IDENTIFIER,
-                APPLY_FIRST_GROUP_KEY_OCCURRENCE_POSITION_QUOTED_IDENTIFIER,
-            ],
+            col_names,
         ),
         input_types=[PandasDataFrameType(col_types=input_types)],
         # We have to specify the local pandas package so that the UDF's pandas